test: Use bats-assert for testing
Description:
------------

`bats-assert' is a helper library providing common assertions for Bats.
On failure, the assertions display helpful details to speed up
debugging. This library significantly increases runtime of the test
suite (see the benchmark below). However, the convenient debug output
and the time it saves greatly outweighs the overhead.

Benchmark:
----------

The benchmark was run on an old Dell Inspiron Mini 9 netbook running
GNU/Linux.

  - CPU: Intel Atom N270 (1.6GHz, dual-core, 32-bit, 512KiB cache)
  - RAM: 1x2GiB DDR2 533MHz
  - OS: Arch Linux
  - Kernel: 4.4.1-2-ARCH
  - Bash: 4.3.42(1)-release
  - Bats: @955309a (after 0.4)

The entire repository was copied into `tmpfs' to avoid disk access. The
runtime of `make check' was measured with the POSIX `time' utility, nine
times back-to-back. The machine was rebooted between testing the
original and the library enhanced version.

Using the library, the runtime shot up by an avery of 135%.

             |  mean  | stan.dev.
    ---------+--------+----------
    original |  99.44 |     1.90
    library  | 233.67 |     1.83

Raw data:
---------

             |  1  |  2  |  3  |  4  |  5  |  6  |  7  |  8  |  9
    ---------+-----+-----+-----+-----+-----+-----+-----+-----+-----
    original | 104    99   100   100    99   100    98    98    97
    library  | 238   234   232   231   234   234   233   234   233